Rename transformer recipe to mt
