+++ src/fairseq2/recipes/__init__.py
@@ -17,7 +17,7 @@ from fairseq2.recipes.assets import _setup_asset_cli
 from fairseq2.recipes.llama import _setup_llama_cli
 from fairseq2.recipes.lm import _setup_lm_cli
 from fairseq2.recipes.logging import setup_basic_logging
-from fairseq2.recipes.transformer import _setup_transformer_cli
+from fairseq2.recipes.mt import _setup_mt_cli
 from fairseq2.recipes.utils.log import exception_logger
 from fairseq2.recipes.wav2vec2.asr import _setup_wav2vec2_asr_cli
 
@@ -49,7 +49,7 @@ def _setup_cli(cli: Cli) -> None:
     _setup_asset_cli(cli)
     _setup_lm_cli(cli)
     _setup_llama_cli(cli)
-    _setup_transformer_cli(cli)
+    _setup_mt_cli(cli)
     _setup_wav2vec2_asr_cli(cli)
 
     # Set up 3rd party CLI extensions.
