+++ src/fairseq2/recipes/mt/__init__.py
@@ -5,28 +5,20 @@
 # LICENSE file in the root directory of this source tree.
 
 from fairseq2.recipes.cli import Cli, RecipeCommandHandler
-from fairseq2.recipes.transformer.eval import (
-    load_transformer_evaluator,
-    transformer_eval_presets,
-)
-from fairseq2.recipes.transformer.translate import (
-    load_text_translator,
-    text_translate_presets,
-)
+from fairseq2.recipes.mt.eval import load_mt_evaluator, mt_eval_presets
+from fairseq2.recipes.mt.translate import load_text_translator, text_translate_presets
 from fairseq2.recipes.utils.sweep import default_sweep_tagger
 
 
-def _setup_transformer_cli(cli: Cli) -> None:
+def _setup_mt_cli(cli: Cli) -> None:
     default_sweep_tagger.extend_allow_set("source_lang", "target_lang")
 
-    group = cli.add_group(
-        "transformer", help="Transformer-based machine translation recipes"
-    )
+    group = cli.add_group("mt", help="machine translation recipes")
 
     # Eval
     eval_handler = RecipeCommandHandler(
-        loader=load_transformer_evaluator,
-        preset_configs=transformer_eval_presets,
+        loader=load_mt_evaluator,
+        preset_configs=mt_eval_presets,
         default_preset="nllb_dense_600m",
     )
